+++ src/commands/BulkMuteCommand.ts
@@ -4,7 +4,7 @@ import {
   InteractionResponseType,
   SlashCommandInteraction,
 } from "harmony";
-import Command from "../structures/Command.ts";
+import Command from "./Command.ts";
 
 /**
  * 実行者が参加しているボイスチャンネルのメンバー全員をミュートするコマンド
